rename NftsListDisplayMode
