+++ svelte/components/Nft/NftsListDisplayMode.svelte
